+++ src/main/java/kitchenpos/application/MenuGroupService.java
@@ -1,9 +1,9 @@
 package kitchenpos.application;
 
 import java.util.List;
+import kitchenpos.application.dto.MenuGroupCreationRequest;
 import kitchenpos.dao.MenuGroupRepository;
 import kitchenpos.domain.MenuGroup;
-import kitchenpos.ui.dto.MenuGroupCreationRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
