+++ src/main/java/kitchenpos/application/dto/MenuGroupCreationRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.application.dto;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
