+++ src/main/java/kitchenpos/application/dto/MenuGroupResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.application.dto;
 
 import kitchenpos.domain.MenuGroup;
