+++ src/main/java/kitchenpos/ui/MenuGroupRestController.java
@@ -4,9 +4,9 @@ import java.net.URI;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.application.MenuGroupService;
+import kitchenpos.application.dto.MenuGroupCreationRequest;
+import kitchenpos.application.dto.MenuGroupResponse;
 import kitchenpos.domain.MenuGroup;
-import kitchenpos.ui.dto.MenuGroupCreationRequest;
-import kitchenpos.ui.dto.MenuGroupResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
