+++ src/main/java/kitchenpos/ui/MenuRestController.java
@@ -3,8 +3,8 @@ package kitchenpos.ui;
 import java.net.URI;
 import java.util.List;
 import kitchenpos.application.MenuService;
+import kitchenpos.application.dto.MenuCreationRequest;
 import kitchenpos.domain.Menu;
-import kitchenpos.ui.dto.MenuCreationRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
