+++ src/test/java/kitchenpos/ui/MenuRestControllerTest.java
@@ -13,10 +13,10 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import java.math.BigDecimal;
 import java.util.List;
 import kitchenpos.application.MenuService;
+import kitchenpos.application.dto.MenuCreationRequest;
+import kitchenpos.application.dto.MenuProductWithQuantityRequest;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.Price;
-import kitchenpos.ui.dto.MenuCreationRequest;
-import kitchenpos.ui.dto.MenuProductWithQuantityRequest;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
